解决删除关联问题

+++ test-system/system-web/src/main/java/com/njust/controller/question/QuestionTypeController.java
@@ -4,6 +4,7 @@ import com.njust.aspectj.annotation.Log;
 import com.njust.aspectj.enums.BusinessType;
 import com.njust.controller.BaseController;
 import com.njust.dto.QuestionTypeDto;
+import com.njust.service.QuestionService;
 import com.njust.service.QuestionTypeService;
 import com.njust.utils.ShiroSecurityUtils;
 import com.njust.vo.AjaxResult;
@@ -27,6 +28,8 @@ public class QuestionTypeController extends BaseController {
     @Reference
     private QuestionTypeService questionTypeService;
 
+    @Reference
+    private QuestionService questionService;
 
 
     /**
@@ -72,7 +75,21 @@ public class QuestionTypeController extends BaseController {
     @DeleteMapping("deleteQuestionTypeByIds/{questionTypeIds}")
     @Log(title = "删除试题类型",businessType = BusinessType.DELETE)
     public AjaxResult deleteQuestionTypeByIds(@PathVariable @Validated @NotEmpty(message = "要删除的ID不能为空") Long[] questionTypeIds) {
-        return AjaxResult.toAjax(this.questionTypeService.deleteQuestionTypeByIds(questionTypeIds));
+        Integer questionCount;
+        Boolean count = false;
+        for (int i = 0; i < questionTypeIds.length; i++) {
+            questionCount = this.questionService.seleceQuestionTypeCount(questionTypeIds[i]);
+            if (questionCount != 0) {
+                count = true;
+                break;
+            }
+        }
+        if (count) {
+            return new AjaxResult(201, "类型已绑定试题，无法删除");
+        } else {
+            return AjaxResult.toAjax(this.questionTypeService.deleteQuestionTypeByIds(questionTypeIds));
+        }
+
     }
     /**
      * 查询所有类型
